Changed int? calories to double? + fixed a paging issue

+++ DietCalculatorSystem/Data/Models/Diet.cs
@@ -16,19 +16,19 @@ namespace DietCalculatorSystem.Data.Models
         public double TotalCarbohydrates { get; set; }
 
         //Breakfast
-        public int BreakfastCalories { get; set; }
+        public double BreakfastCalories { get; set; }
         public double BreakfastProteins { get; set; }
         public double BreakfastFats { get; set; }
         public double BreakfastCarbohydrates { get; set; }
 
         //Lunch
-        public int LunchCalories { get; set; }
+        public double LunchCalories { get; set; }
         public double LunchProteins { get; set; }
         public double LunchFats { get; set; }
         public double LunchCarbohydrates { get; set; }
 
         //Dinner
-        public int DinnerCalories { get; set; }
+        public double DinnerCalories { get; set; }
         public double DinnerProteins { get; set; }
         public double DinnerFats { get; set; }
         public double DinnerCarbohydrates { get; set; }
